BE/requirements.txt
fastapi
uvicorn
SpeechRecognition
pydub
youtube-transcript-api
openai-whisper
nltk
yt_dlp
transformers